MAINT: rename private module from _classifier to _bagging
